include <cmath>

+++ src/LayerSoftmax.cpp
@@ -10,6 +10,8 @@
 
 #include "Activation.h"
 
+#include <cmath>
+
 ///////////////////////////////////////////////////////////////////////////////
 LayerSoftmax::LayerSoftmax():
     Layer(0,0,"Softmax")
@@ -51,7 +53,7 @@ void LayerSoftmax::backpropagation(const MatrixFloat &mInput,const MatrixFloat &
 		float s = S.array().exp().sum();
 		for (int c = 0; c < S.cols(); c++)
 		{
-			float expx = exp(S(c));
+			float expx = expf(S(c));
 			mGradientIn(r, c) = mGradientOut(r, c)*(expx*(s-expx)) / (s*s);
 		}
 	}
